requirements.txt
asgiref==3.8.1
boto3==1.35.25
botocore==1.35.25
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.8
jmespath==1.0.1
numpy==2.1.0
packaging==24.1
pandas==2.2.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
psycopg2==2.9.10
SQLAlchemy==2.0.23